Added more tests for our default_page events.
I was inclined to remove the check_default_page event handler,
because the current tests passed just fine without that handler.
But it is needed after all.
Note that getDefaultPage asks for the default_page property
and returns None when it is not there.
So in these new tests we check that the property itself was emptied.