Removed Py2-only test from Products/CMFDynamicViewFTI/tests/test_fti.py
And remove six from the rest of the code.

According to a comment, originally on Py 3 it yielded a TypeError: 'ReplaceableWrapper' object is not callable.
On Plone 6 I do not see this.  But the two tests in here still fail with a NotFound.
When I try to recreate the situation of this test setup manually in Plone 6, it works fine:
create a folder and document, give them both an empty layout property, and try to view them.
So at least for a Dexterity FTI it seems to go fine.