fix: run test with tox
Turns out that when `plone.testing` installs distributions from the
`Products` namespace, it enforces that all of them have to be in the
same folder (which kind of make sense) but `tox` has a different idea on
that :-/

`tox` installs all dependencies but the development package is kept
outside that structure, and thus the `plone.testing` condition is not
met.

With this patch we check if it is the `tox` case and force the install.